refactor: Update Gulp
- Re-order tasks into "task" and "helper" namespaces.
- Add a task-status and task-ports which print whether a Gulp
instance is running and the ports being used by Gulp, respectively.
- Add task-tohtml to convert any MarkDown file to its HTML
equivalent.
- Add task-clear Gulp task to clear the contents of any
./gulp/.gulp* file.
- Improve the default task to also be able to handle a --stop flag.
When provided the Gulp instance will stop.
- Instead of using 2 browser-sync servers use a single server for
both the app development and the documentation. This makes things
cleaner and easier to maintain. The only drawback is that all bs
tabs are now reloaded.
- Re-work the task-watch Gulp task. Changes make sure everything is done
asynchronously now.
- Add more browser names to the possible browser names list array to
cover Ubuntu and MacOS in config.json.
- Implement branch switch fail-safe. Unintended consequences can occur
when Gulp is running and a 'git checkout' command is run. Therefore,
when changing branches Gulp will exit the process and notify the user.
Simple restart Gulp after the branch switch.
- Other cleanup and task improvements.
- Improve documentation where needed.