style: Change to prettier formatter
- Change JS formatter to Prettier from JsBeautify.
	- Prettify project files with Prettier.
- Implement the ".ig." sub-file extension. It stands for ignore.
Any file with this sub-extension will be ignored when the pretty
Gulp task is run. These files have to be ignored for one reason
or another. In this case the files don't contain properly finished
JS code so the Prettier formatter will throw an error and halt the
Gulp pipe.
- Add a --exact flag to the files Gulp task. When this flag is
provided fuzzy search will be switched for an indexOf string
search. Sometimes this is needed over a fuzzy search.