refactor: Update gulpfile.main.js
- Replace CSSComb with Perfectionist. CSSComb has some issues and
Perfectionist seems to be the alternative to use. It can also be
used with PostCSS and custom settings can be applied as well.
- Remove purify.js code from gulpfile.main.js for the time being.
The current purifiers have their issues and is something that might
need more tweaking to make it work as intended. Therefore, for now
it's best to remove the code and save this feature for later.
- Beef up the pretty Gulp task by adding more flags. These flags
make the task more flexible and useful.