fix: Gulp pretty task destination path issue
Previously when a glob was being provided via --glob the task was
would not dump the file back to its original location. The files,
regardless of their initial location, would all end up in the root.
This is now resolved and files, once prettified, are dumped back
to their location of origin. This was resolved by simply using the
"base" option with gulp.src().