fix: gulp-task-doc indentation hack
Add an indentation hack for "long" comments. Long comments spacing
for some reason gets trimmed. This "hack"/modification seems to be the
simplest solution at the moment. In short, the solution will indent
it the same amount as the former lines. The "=>" sets the start point
and each following line should start with "/". These characters will
be removed and not be shown in final terminal input.