fix: Update paths object
Assign the correct __dirname and cwd paths to the object before
expanding the paths. Moving the expand_paths function to utils.js
was grabbing the wrong __dirname and cwd. To ensure the correct ones
are used again, they are assigned to the path object after grabbing
the needed paths object.