feat: Add a cli_highlight function
As the helpers/pretty.js file was removed from gulpfile.setup.js
in favor of running the pretty task from gulpfile.main.js using
node-cmd from within gulpfile.setup.js, the returned output from
the command is only a simple string with any highlighting removed.
This function adds the highlighting back to the string to be
consistent with all the other task outputs.