fix: Add preventive file filter to pretty.js (--quick/--staged)
When using the --quick/--staged flags they will add the modified
or staged files provided via the Git command. This poses an issue
as it will provided files that maybe should not be passed into the
pretty task like image and Markdown files, for example. Therefore,
this commit filters out all files but the relevant files before
prettifying.