fix: Update .__internal.json cleanup
When a Gulp process is started, process information is stored in an
internal file called .__internal.json. The stored information is also
cleared when the Gulp instance is closed. However in the case that
an error makes the Gulp process close the file might not be cleared
properly. This causes an issue: when the file still contains the
previous Gulp process's information a new Gulp instance cannot be
started. One has to manually clear the file. To make this better and
programmatic a new way of handling the issue has been implemented.
The new method checks whether the old process is still in use. When
the process does not exist a new Gulp process is allowed to be made.