feat: Add a --linkify flag option to tohtml task
The linkify flag tries to turn relative URLs to an absolute HTTP
URL. This is only done for previewing purposes as the .html file's
destination when converted is different than from where the file
is originally located (the source .md file). This difference breaks
relative link paths so this feature finds all relative paths and
converts them to absolute HTTP paths to work with its new
destination.