refactor: Implement a consistent reporter for linters
- Remove csslint-stylish and jshint-stylish.
- Update gulp-jshint to from 2.0.4 to 2.1.0. Old version had
issues with template literal strings and variables saying the
variable(s) was not used when placed in a template string. New
version has this issue resolved.
- Implement a custom and single reporter for plugins gulp-csslint,
gulp-htmllint, and gulp-jshint. Before, the three plugins gave
varying outputs as well as had different ways implementation. Now
they all work the same way and give consistent output. This is
better for future maintainability.